requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
scikit-learn==0.24.2
matplotlib==3.4.3
RPi.GPIO==0.7.0